Move INodeRegistry to raftio/registry
